pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pyquality"
version = "0.1.0"
description = "A highly user-friendly, kid-friendly automation framework using Chrome DevTools Protocol (CDP)."
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
authors = [
  { name="Author Name" }
]
dependencies = [
    "websocket-client>=1.6.1",
    "requests>=2.31.0"
]

[project.urls]
"Homepage" = "https://github.com/yourusername/pyquality"
"Bug Tracker" = "https://github.com/yourusername/pyquality/issues"
